args: Disallow -noconf
ArgsManager::GetConfigFilePath() asserts that a path is always set. Explicitly negating it is therefore invalid.

Previously, -noconf would result in an ifstream being opened to the ".bitcoin"-directory (not a file) in ArgsManager::ReadConfigFiles(), also circumventing the assert in aforementioned GetConfigFilePath().

(-conf="" is still allowed to reset to the default bitcoin.conf-file location).